much done

+++ apps/api/src/arguments/LocationInput.ts
@@ -4,6 +4,6 @@ export default inputObjectType({
 	name: 'LocationInput',
 	definition: (t) => {
 		t.field({ name: 'location', type: 'Location' })
-		t.nullable.int('range')
+		t.nullable.float('range')
 	},
 })
